Replace array class code with ndarray class plus copy
While my previous fix for the memory leak problem fixed that issue,
other code in generic_io.py and its tests expects that the returned
object be an ndarray. Since using the frombuffer call followed by a
copy to break the reference achieves the same result as using array,
I've replaced my pervious change to MemryIO with this new change.