Fix version error message
Asdf throws an error message when it fails to read the version map
file. The error message includes the version being searched for in the
form of an undefined variable, version_string. I added a line defining
this variable so that the exception doesn't throw another exception.

Please note that the most likely reason for the version map file not
to be found is that the asdf-standard package has not been installed,
as that is where the file is found in the distribution.