Force new blocks to be contiguous
The code in asdf assumes a block will be a contiguous block of
memory. But recent versions of numpy have used views more heavily and
now there are cases when an object instantiated as a bock is not
contiguous, which leads to errors when the block is used. The modified
code calls numpy.ascontiguousarray to make sure the block memory is
contiguous.

This is a revised version of the branch, as several other irrelevant
changes were put in the previous verion of the branch.