Workaround in unit test for lack of CartRepr comparator...
CartesianRepresentation in astropy does not define equality in a way
that is useful for our unit tests, so we have to implement the
comparison ourselves. There may be a way to do it that removes the
dependency on astropy from the test helpers but this was the simplest
solution for now.