First pass at backwards compatibilty for tag class...
This is not meant to be the last word for backwards compatibility. In
fact, it's really only the first word. There's a lot more work to be
done and more thought to be given to the mechanisms that are required to
support this kind of behavior. Also, this particular case isn't even all
that critical--it just serves as a potentially useful example.