Add unit test for tag implementation newer than schema read...
This is an important case that is not currently handled properly.
Currently if an exact version match is not found for a schema version,
the "best match" is used. However, the tag implementation for this "best
match" may not be compatible with the given schema version, which will
result in an unhandled error. This needs to be handled properly.