Custom tag class in unit test shouldn't inherit AsdfType...
All classes that inherit AsdfType are automatically added to the
internal list of built-in types. In this case, adding a type that was
defined within a unit test to the list of built-ins caused problems for
the schema validation tests. This commit is really just a work-around
for now. In the long term, there should be another class that can be
inherited by both built-in and user-defined types.