WIP: moved version compat logic into yamlutils...
This seems like the most logical place to make a decision about whether
to write a custom data structure or possible or just return a raw Python
object. Added a new method to AsdfTypes `version_is_supported` to
determine if a particular tag version can be handled by the class. It's
mostly just a stub for now but logic will be fleshed out soon.