Extension classes now have `supported_version` field...
Custom extension classes can now declare explicit support for various
schema versions using the `supported_version` field. If this field is
provided by an extension class, it will be used to determine whether that
class can be used to convert a particular version of a schema to/from a
custom class. If the field is not provided, or if a particular schema
version is not listed, the extension class will simply return raw Python
data structures representing the schema.