Prototype of way to have multiple versions in one tag class...
This update allows a single extension class to support multiple versions
of the same schema tag. There might be a better way to do it in the long
run, but this way appears to work, at least for built-in classes. It
would require a different mechanism to be used for user-defined classes,
which is problematic.