Refactor to handle versioned tag classes in ExtensionList creation...
In order to correctly implement multiple supported versions of tag
classes for all extensions and not just ASDF built-ins, the logic for
adding versioned extension classes was moved from the AsdfType metaclass
constructor to the constructor for AsdfExtensionList. This allows both
AsdfType and UserType sibling version classes to be properly handled.
The sibling version classes themselves are now created within the
ExtensionType constructor and are stored as property of the main version
class.