YAML validation warning whenever not latest version...
This is to make sure that warnings occur even when there are tag classes
for a particular tag, as long as that tag does not represent the latest
known version of that tag. The purpose of this commit is to prevent
`supported_versions` from having an affect on YAML validation warnings.

This also refactors the warning handling logic to consolidate and reuse
code.