Update mechanism used to handle missing tag requirements...
Instead of directly issuing a warning passing the raw node back in this
case, raise an exception that will be handled by
tagged_tree_to_custom_tree, which will then use the existing warning
mechanism to indicate a failure to convert the tree. This is much
cleaner since it reuses mechanisms that were recently improved; and it
avoids some issues that occurred when running tests against older
versions of astropy.