Add top-level conftest to avoid conflict with astropy...
Since parts of ASDF depend on astropy, and now parts of astropy depend
on ASDF, we need to be careful about circular dependencies. In this
case, having the function enable_deprecations_as_exceptions in a
conftest file within the ASDF source tree caused problems when running
the astropy test suite. Moving this function to a conftest outside the
source tree appears to have solved the problem.