when memmapping, memmap entire file
Previous to the PR when memmapping each memmapped array would
consume one open file. For files with large numbers of arrays
this increases the liklihood of hitting the maximum number of
open files for a single process and also drastically slows down
file open times.

This PR memmaps the entire file (when memmapping is enabled)
and provides the memmap (a python mmap.mmap object from the
standard library) to each memmapped array (via a call to
ndarray.__new__).

File updates (asdf.update) complicates the handling of this
whole file memmapping as it can result in changes to the file
on disk which require regeneration of the memmap. The test
test_block_data_change checks for this regeneration.