add NDArrayType support to get_array_base and fits block finding
NDArrayType fails checks to isinstance(arr, ndarray) as it is not a
subclass of ndarray. This prevented util.get_array_base from finding
the correct base array which contributed to failures to link hdu
items to arrays within the tree (leading to duplication of data).

The EmbeddedBlockManager in fits_embed also skipped checking for
links between NDArrayType and hdu items which also broke the link
between hdu items and arrays within the tree. This PR removes the
skip.

fixes issue #1232