support integer-1.1 and ndarray-1.1
fixes an issue with integer where InterType instead of cls
was used during class instantiation which resulted in objects
being created with the incorrect version. Prior code (that only
had one integer version, 1.0) did not suffer from issues because
of this bug.

fixes #1245 by removing the use of super in NDArrayType and adding
an exception to check for incompatible use of super in classes of
type ExtensionTypeMeta.