don't limit IntegerType _value_cache per ctx
_value_cache is used to reuse arrays (and eventually blocks) when
multiple IntegerType instances share the same value (and can reuse
the same block). A separate cache was used for each ctx (AsdfFile).
However, this cache stores ndarrays which can be shared across
files as the relationship between arrays and blocks is managed
by separate block managers for each file allowing each file to
create distinct blocks. This preserves the main use of the cache
which is to return an array with the same id for IntegerTypes
of identical value to allow the block manager to reuse blocks.