change import in fits_embed tests
to allow any ordering of tests and to check that a warning
is emitted each time asdf.fits_embed is imported the module
is removed before importing. This is needed because the first
import of asdf.fits_embed will be cached in sys.modules.
Every subsequent import will used the stored version and not
emit a warning. However importing portions of the module
(such as from asdf.fits_embed import AsdfInFits) will cause
this approach to fail as the del call to sys.modules will
not fully remove the module and any subsequent import
of asdf.fits_embed will not emit a new warning