add upper pins to dependencies
these are added to packages using the following strategy:
- development version (<1.0) are left unpinned as it seems overly
  conservative to pin these to a minor version and as they're <1.0
  there is little guarantee that a minor version won't introduce
  a breaking change. Packages effected are:
    asdf-transform-schemas
    asdf-unit-schemas
    jmespath
- packages that seem to follow semantic versioning were pinned
  to the current working major version. These are:
    asdf-standard
    importlib-metadata
    importlib_resources
    jsonschema
    pyyaml
    semantic_version
- packages that don't follow semantic versioning were pinned
  to the current working version. These are:
    numpy
    packaging

previously there was no upper pin for numpy except
for python 3.8 (which is incompatible with numpy 1.25)
since the upper pins are being set here before 1.25 is
released this would create a duplicate requirements line
(since we aren't yet sure what changes will be included
in 1.25). The duplication was removed and a note added
in case the numpy upper pin is removed (or relaxed).