free tree on AsdfFile.close
With these changes, when an AsdfFile object is closed
the tree (which already could not be accessed because of the
tree property gatekeeping) is set to a blank AsdfObject.

This allows the garbage collector to clean up objects
referenced in the tree even though the AsdfFile object
does not fall out of scope.

Fixes #1558