convert numpy scalars to python types before yaml encoding
nep51 changes numpy scalars modifying the repr.

this change conflicts with represent_float/int (which uses
repr) and with numpy 2.0 (which implements nep51) scalars
are being encoded as, for example, 'np.float64(3.14)'
instead of '3.14'. To work around this change, convert the
values to builtin python types before passing them to
pyyaml for encoding.