remove auto_inline test
this test was previously never run because of
the odd behavior of pytest when filterwarnings
is provided an exception class (instead of a string).
This results in the exception class being decorated
with the mark, then the function being provided as the
first argument to the class (as the exception class
is used as the decorator). Because the exception class
does not start with "Test" the test is ignored during
collection.