re-add check for >1 tag without select_tag in ConverterProxy
turn error into warning for Converter subclasses that support >1 tag

Previously this was allowed because of the default select_tag
provided by Converter. Removing this default breaks at least
one downstream package (gwcs) which relied on this bug. To help
ease the transitiono to a Converter that doesn't provide
select_tag (and fixes the bug caused by this) this commit
changes the error to a warning for any convert subclasses
that support >1 tag but do not implement select_tag. This
more closely matches the previous behavior but introduces
a warning to help packages relying on this bug to know they
should be updated.

add warning filter for gwcs extension bug