index converters by type when no converter is found
This changes how converters with class paths/names
as types are handled by asdf.

Prior to this commit, when converting a class instance
(foo) of class (Foo) the class path of the instance
was inspected with ``get_class_name`` and the
``_converters_by_type`` index was searched using
the class name. This index contained both classes
and class paths (as converters could use both/either)
with classes being preferred. This creates issues
when a module has a different class path from the
typical 'public' path for the class as the module
might move the private implementation without changing
the 'public' path with the expectation that this will
not break downstream code. For converters that use
class paths these types of moves will break the converter.

This commit changes the handling of class paths used
for converters so that the 'public' path can be used.
If a converter is not found for a class in
``_converters_by_type`` (which now only contains classes
as keys) then ``_converters_by_class_path`` is indexed
by checking what classes referred to by the paths
are already imported (to avoid importing every class
supported by every extension). If the class is imported
it is added to ``_converters_by_type`` and removed from
``_converters_by_class_path``.