allow memmapped arrays to keep mmap open
asdf uses a mmap to memory map files to produce np.memmap
arrays (when copy_arrays is True, the current default).
np.memmap does not check if the backing mmap is closed before
accessing array data and asdf closes the mmap when the asdf
file is closed. This results in segfaults (see issue 1334).

This commit changes when happens to the mmap when the asdf
file is closed. asdf will no longer force closure of the mmap.
This will result in an open file pointer if any array still
references the mmap (the mmap will likely be closed by the
garbage collector when all arrays are collected which matches
the behavior of np.memmap).