index converters after discovery
this fixes an issue with extension order and class path.

If extension 1 is found first, and registers a converter using
a class path (with a module that hasn't been imported). Then,
extension 2 registers the same class but uses the class (not the
path). Prior to this commit (in this branch) asdf would use
extension 2 for the type (because the class path registered
with extension 1 was never indexed). With this commit all
converter class paths/types are recorded and then indexed
(in the order they're registered) fixing this issue.